# docs/openapi.yaml
...
openapi: 3.0.1
info:
  title: SEGS API
  description: 'Secure Email Gateway Suite HTTP API. Authenticate with a stateful JWT (``Authorization:
    Bearer``) or the ``seg_session`` cookie. SCIM 2.0 is at ``/scim/v2``. Interactive ``/docs``
    is not served; the contract is ``docs/openapi.yaml`` and is what API Gateway imports.'
  version: 1.0.0
  x-generated-by: backend.api.openapi
servers:
- url: /
  description: Same origin as the SOC console (CloudFront)
paths:
...
  /api/enforcement:
    get:
      summary: Get Enforcement
      description: Return current runtime enforcement mode.
      operationId: get_enforcement_api_enforcement_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - bearerAuth: []
      - cookieAuth: []
    put:
      summary: Set Enforcement
      description: Set enforcement mode. Detection and monitoring always run; only blocking
        changes.
      operationId: set_enforcement_api_enforcement_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnforcementUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
  /api/campaigns:
    get:
      summary: List Campaigns
      operationId: list_campaigns_api_campaigns_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
  /api/campaigns/by-id:
    get:
      summary: Get Campaign
      operationId: get_campaign_api_campaigns_by_id_get
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: string
          minLength: 4
          maxLength: 40
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
      - cookieAuth: []
...
components:
  schemas:
...
    EnforcementUpdate:
      properties:
        mode:
          type: string
          title: Mode
      type: object
      required:
      - mode
      title: EnforcementUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateful JWT (RFC 6750). Same token as the seg_session cookie; jti is stored
        server-side.
    cookieAuth:
      type: apiKey
      in: cookie
      name: seg_session
      description: Session cookie set at login. Same credential as Bearer.
    scimBearer:
      type: http
      scheme: bearer
      description: SCIM 2.0 bearer (SEG_SCIM_BEARER_TOKEN) or an admin JWT.